Delt opp SketchHeader i komponenter, fikset knappelogikk
