+++ TrafikkApp/components/sketchComponents/SketchSizeMenu.js
@@ -1,0 +1,13 @@
+// const PencilSizePopup = () => {
+
+//     return (
+//         <View>
+//             <Menu renderer={Popover} rendererProps={{ preferredPlacement: 'bottom' }}>
+//             <MenuTrigger><Icon name={'circle'} solid size={32} color={currentColorSetup.color}/></MenuTrigger>
+//             <MenuOptions>
+//                 <MenuOption style={styles.colorMenu} ><ColorPalette value={currentColorSetup.color} onValueChange={onPaletteValueChange} colors={colors} numberOfRows={2} containerWidth={200} /></MenuOption>
+//             </MenuOptions>
+//             </Menu>
+//         </View>
+//     )
+// };
